requirements.txt
Pillow~=9.2.0
numpy~=1.23.4